refactor(copyright): split tree walk modules
